+++ examples/server_sent_events/src/main.rs
@@ -1,38 +1,17 @@
 use std::time::Duration;
 
 use picoserve::{
-    response::{self, StatusCode},
+    response::{self, ErrorWithStatusCode, StatusCode},
     routing::{get, get_service, post},
-    ResponseSent,
 };
 
+#[derive(Debug, thiserror::Error, ErrorWithStatusCode)]
+#[status_code(BAD_REQUEST)]
 enum NewMessageRejection {
+    #[error("Read Error")]
     ReadError,
+    #[error("Body is not UTF-8: {0}")]
     NotUtf8(std::str::Utf8Error),
-}
-
-impl response::IntoResponse for NewMessageRejection {
-    async fn write_to<R: picoserve::io::Read, W: response::ResponseWriter<Error = R::Error>>(
-        self,
-        connection: response::Connection<'_, R>,
-        response_writer: W,
-    ) -> Result<ResponseSent, W::Error> {
-        match self {
-            NewMessageRejection::ReadError => {
-                (StatusCode::BAD_REQUEST, "Read Error")
-                    .write_to(connection, response_writer)
-                    .await
-            }
-            NewMessageRejection::NotUtf8(err) => {
-                (
-                    StatusCode::BAD_REQUEST,
-                    format_args!("Body is not UTF-8: {err}\n"),
-                )
-                    .write_to(connection, response_writer)
-                    .await
-            }
-        }
-    }
 }
 
 struct NewMessage(String);